Add new properties to CarDetailDto
This commit adds new properties to CarDetailDto. It also changes the variable type of the ModelYear property of the CarDetailDto and Car entities.

It also adds an Expression argument to the GetCarDetails method in the ICarDal interface. With this change, different methods can be created in Business classes.

It also fixes the typo (c.BrandId -> c.ColorId) in DataAccess/Concrete/EntityFramework/EfCarDal:22.