Fix wwwroot path
This commit, fixes the wwwroot path (WebAPI/Uploads/wwwroot -> WebAPI/wwwroot).

It also fixes currentDirectory path in FileHelper. Moves the default.jpg file to the new wwwroot folder.

It also adds the relevant code to the Startup.cs file to be able to publish the migrated wwwroot folder with the API.