Fix Rental Logic
This commit, fixes the car rental logic in general.

It also adds new properties to RentalDetailDto.

It also renames the GetRentalsDto method in the IRentalDal interface and in the EfRentalDal class to GetRentalsDetails and adds an expression argument to this method.

It also adds a method to the project that retrieves all the rental details by customer id.