Add Payment System
This commit adds a payment system to the project.

It adds the Rent method to IRentalService and RentalManager, which allows renting a car with a payment system. Provides the Rent method in the API.

It also changes the name of a method in RentalManager in the whole project.

It also fixes the WebAPI.csproj file in addition to the 25b784e commit.